+++ idea/testData/refactoring/rename/renameKotlinEnumConstByJavaRef/before/EnumJavaRef.java
@@ -1,0 +1,5 @@
+public class EnumJavaRef {
+    public void use(EnumTarget p) {
+        System.out.println(p == EnumTarget./*rename*/RED);
+    }
+}