+++ idea/testData/refactoring/rename/renameKotlinPropertyWithJvmField/before/RenameKotlinPropertyWithJvmField.kt
@@ -1,0 +1,5 @@
+package testing.rename
+
+public open class C {
+    @JvmStatic var <caret>first = 1
+}